make file - optional in  endpoint with addding the file

+++ backend/models/progress.py
@@ -77,5 +77,5 @@ class CourseProgress(BaseModel):
 
     def completion_percentage(self):
         if self.total_assignments == 0:
-            return 0
-        return round((self.completed_assignments / self.total_assignments) * 100, 2)
+            return 0.0
+        return float(round((self.completed_assignments / self.total_assignments) * 100, 2))
